% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.get_param_names}
\alias{.get_param_names}
\title{Get human-readable parameter names from a fit object}
\usage{
.get_param_names(fit)
}
\arguments{
\item{fit}{A \code{seroreconstruct_fit} or derived object.}
}
\value{
Character vector of parameter names.
}
\description{
Get human-readable parameter names from a fit object
}
\keyword{internal}
