% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{.new_seroreconstruct_fit}
\alias{.new_seroreconstruct_fit}
\title{Constructor for seroreconstruct_fit objects}
\usage{
.new_seroreconstruct_fit(
  output,
  n_groups,
  n_individuals,
  n_posterior_samples,
  runtime_secs
)
}
\description{
Constructor for seroreconstruct_fit objects
}
\keyword{internal}
