% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{.new_seroreconstruct_joint}
\alias{.new_seroreconstruct_joint}
\title{Constructor for seroreconstruct_joint objects}
\usage{
.new_seroreconstruct_joint(fit, group_labels, group_sizes, shared, n_groups)
}
\description{
Constructor for seroreconstruct_joint objects
}
\keyword{internal}
