% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{.new_seroreconstruct_multi}
\alias{.new_seroreconstruct_multi}
\title{Constructor for seroreconstruct_multi objects}
\usage{
.new_seroreconstruct_multi(fits, group_labels, group_sizes)
}
\description{
Constructor for seroreconstruct_multi objects
}
\keyword{internal}
