% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.plot_traces}
\alias{.plot_traces}
\title{Plot MCMC trace plots}
\usage{
.plot_traces(mcmc_matrix, nrow, ncol)
}
\arguments{
\item{mcmc_matrix}{Matrix of posterior samples.}

\item{nrow}{Number of rows in the plot layout.}

\item{ncol}{Number of columns in the plot layout.}
}
\description{
Plot MCMC trace plots
}
\keyword{internal}
