% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_simulation_params}
\alias{.validate_simulation_params}
\title{Validate simulation parameters}
\usage{
.validate_simulation_params(para1, para2)
}
\arguments{
\item{para1}{Numeric vector of length 10.}

\item{para2}{Numeric vector of length 20.}
}
\description{
Validate simulation parameters
}
\keyword{internal}
