% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{inputdata}
\alias{inputdata}
\title{Example of input data}
\format{
A data frame with 9 variables, where each row represents an individual:
\describe{
  \item{age_group}{0: children, 1: adults, 2: older adults}
  \item{start_time}{start of follow-up}
  \item{end_time}{end of follow-up}
  \item{time1}{date of first serum collection}
  \item{time2}{date of second serum collection}
  \item{time3}{date of third serum collection}
  \item{HAI_titer_1}{HAI titer for first serum collection}
  \item{HAI_titer_2}{HAI titer for second serum collection}
  \item{HAI_titer3}{HAI titer for third serum collection}
}
}
\usage{
data(inputdata)
}
\description{
This is an example of the input data used in the \code{seroreconstruct} function. This data frame illustrates the format of the input data.
}
\concept{inputdata}
\keyword{datasets}
