% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_function.R
\name{output_model_estimate}
\alias{output_model_estimate}
\title{Extract the model estimates from the fitted MCMC}
\usage{
output_model_estimate(fitted_MCMC, period)
}
\arguments{
\item{fitted_MCMC}{The fitted MCMC}

\item{period}{A vector indicate the start and the end of a season to compute the infection probabilities. If empty, the start and the end of a season was assumed to be the minimum of the start of follow-up and the maximum of the end of follow-up among individuals.}
}
\value{
A data frame that store the estimates from a fitted MCMC.
}
\description{
The function to obtain the estimates of the infection probabilities, boosting, waning and measurement error
}
\examples{
\donttest{
a1 <- sero_reconstruct(inputdata, flu_activity,
                        n_iteration = 2000, burnin = 1000, thinning = 1)
fitted_result <- output_model_estimate(a1)
}
}
