% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{para2}
\alias{para2}
\title{Example of parameter vector for the baseline HAI titer for the main model}
\format{
A numeric vector with 20 elements (for a single-season model, \code{S = 1}).
  The general length is \code{20 * S} where \code{S} is the number of seasons.
\describe{
  \item{Elements 1--10}{probability that the HAI titer is 0--9 for children}
  \item{Elements 11--20}{probability that the HAI titer is 0--9 for adults}
}
For multi-season models, this pattern repeats for each season.
}
\usage{
data(para2)
}
\description{
This is an example of the parameter vector for the baseline HAI titer for the main model used in the \code{seroreconstruct} function. This data frame specifies the format of the parameter vector for the baseline HAI titer for the main model.
}
\seealso{
Other example_data: 
\code{\link{flu_activity}},
\code{\link{para1}}
}
\concept{example_data}
\keyword{datasets}
