% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{print.seroreconstruct_fit}
\alias{print.seroreconstruct_fit}
\title{Print method for seroreconstruct_fit}
\usage{
\method{print}{seroreconstruct_fit}(x, ...)
}
\arguments{
\item{x}{A \code{seroreconstruct_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for seroreconstruct_fit
}
