% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{print.seroreconstruct_joint}
\alias{print.seroreconstruct_joint}
\title{Print method for seroreconstruct_joint}
\usage{
\method{print}{seroreconstruct_joint}(x, ...)
}
\arguments{
\item{x}{A \code{seroreconstruct_joint} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for seroreconstruct_joint
}
