% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{print.seroreconstruct_multi}
\alias{print.seroreconstruct_multi}
\title{Print method for seroreconstruct_multi}
\usage{
\method{print}{seroreconstruct_multi}(x, ...)
}
\arguments{
\item{x}{A \code{seroreconstruct_multi} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for seroreconstruct_multi
}
