% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{print.summary.seroreconstruct_fit}
\alias{print.summary.seroreconstruct_fit}
\title{Print method for summary.seroreconstruct_fit}
\usage{
\method{print}{summary.seroreconstruct_fit}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{summary.seroreconstruct_fit} object.}

\item{digits}{Number of decimal places for rounding. Default 2.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for summary.seroreconstruct_fit
}
