% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{print.summary.seroreconstruct_joint}
\alias{print.summary.seroreconstruct_joint}
\title{Print method for summary.seroreconstruct_joint}
\usage{
\method{print}{summary.seroreconstruct_joint}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{summary.seroreconstruct_joint} object.}

\item{digits}{Number of decimal places for rounding. Default 2.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for summary.seroreconstruct_joint
}
