% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct-package.R
\docType{package}
\name{seroreconstruct-package}
\alias{seroreconstruct}
\alias{seroreconstruct-package}
\title{seroreconstruct: Reconstructing Antibody Dynamics to Estimate the Risk of Influenza Virus Infection}
\description{
A Bayesian framework for inferring influenza infection status from serial antibody measurements. Jointly estimates season-specific infection probabilities, antibody boosting and waning after infection, and baseline HAI titer distributions via MCMC. Supports multi-season analysis and subgroup comparisons via a group_by interface. See Tsang et al. (2022) \doi{10.1038/s41467-022-29310-8} for methodological details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timktsang/seroreconstruct}
  \item Report bugs at \url{https://github.com/timktsang/seroreconstruct/issues}
}

}
\author{
\strong{Maintainer}: Tim Tsang \email{timkltsang@gmail.com} (\href{https://orcid.org/0000-0001-5037-6776}{ORCID})

}
\keyword{internal}
