% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_function.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulation of the dataset of the Bayesian model}
\usage{
simulate_data(inputdata, inputILI, para1, para2)
}
\arguments{
\item{inputdata}{The data with the same format that for running MCMC, in dataframe format.}

\item{inputILI}{The data for influenza activity used in the inference. The row number should match with the date in the inputdata.}

\item{para1}{The parameter vector for the model parameters, in the following format: 1) the parameter value of the random measurement error, 2) the parameter value of the 2-fold error, 3) the boosting in HAI titer after infection for children (in log2 unit), 4) the waning in HAI titer for children (in log2 unit), 5) the boosting in HAI titer after infection for adults (in log2 unit), 6) the waning in HAI titer for adults (in log2 unit), 7) the scale parameter for children, 8) the scale parameter for adults, 9) the scale parameter for older adults, 10) the log of risk ratio of 2-fold increase in baseline HAI titer}

\item{para2}{The parameter vector for the parameter for the HAI titer distribution. The first 10 elements are the probability that the HAI titer is equal to 0-9 for children, and the elements of 11-20 is the probability that the HAI titer is equal to 0-9 for adults.}
}
\value{
A simulated data based on the input parameter vectors, with the format equal to the input data.
}
\description{
The function to simulate the dataset, for validation or other purpose.
}
\examples{
simulated <- simulate_data(inputdata, flu_activity, para1, para2)
}
