% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{summary.seroreconstruct_joint}
\alias{summary.seroreconstruct_joint}
\title{Summary method for seroreconstruct_joint}
\usage{
\method{summary}{seroreconstruct_joint}(object, period, ...)
}
\arguments{
\item{object}{A \code{seroreconstruct_joint} object.}

\item{period}{Optional numeric vector of length 2 specifying the start and end
of a season to compute infection probabilities.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{summary.seroreconstruct_joint} object with element \code{$table}.
}
\description{
Computes shared parameter estimates and per-group infection probabilities
from a joint fit with shared parameters.
}
