% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct_fit.R
\name{summary.seroreconstruct_multi}
\alias{summary.seroreconstruct_multi}
\title{Summary method for seroreconstruct_multi}
\usage{
\method{summary}{seroreconstruct_multi}(object, period, ...)
}
\arguments{
\item{object}{A \code{seroreconstruct_multi} object.}

\item{period}{Optional numeric vector of length 2 specifying the start and end
of a season to compute infection probabilities.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{summary.seroreconstruct_multi} object with element \code{$table}.
}
\description{
Computes estimates for each group and combines into a single table.
}
