% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroreconstruct-package.R
\docType{package}
\name{seroreconstruct-package}
\alias{seroreconstruct}
\alias{seroreconstruct-package}
\title{seroreconstruct: Reconstructing Antibody Dynamics to Estimate the Risk of Influenza Virus Infection}
\description{
A Bayesian model to jointly estimate infection probability in seasons, boosting and waning in antibody after infection, and baseline distribution of HAI titer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timktsang/seroreconstruct}
  \item Report bugs at \url{https://github.com/timktsang/seroreconstruct/issues}
}

}
\author{
\strong{Maintainer}: Tim Tsang \email{timkltsang@gmail.com} (\href{https://orcid.org/0000-0001-5037-6776}{ORCID})

}
\keyword{internal}
